Moved around settings functions
