+++ packages/foam-vscode/src/services/logging.ts
@@ -1,7 +1,10 @@
-import { window, commands, ExtensionContext } from 'vscode';
+import { workspace, window, commands, ExtensionContext } from 'vscode';
 import { IDisposable } from '../core/common/lifecycle';
 import { BaseLogger, ILogger, LogLevel } from '../core/utils/log';
-import { getFoamLoggerLevel } from '../settings';
+
+function getFoamLoggerLevel(): LogLevel {
+  return workspace.getConfiguration('foam.logging').get('level') ?? 'info';
+}
 
 export interface VsCodeLogger extends ILogger, IDisposable {
   show();
