+++ packages/foam-vscode/src/settings.ts
@@ -1,6 +1,5 @@
 import { workspace, GlobPattern } from 'vscode';
 import { uniq } from 'lodash';
-import { LogLevel } from './core/utils/log';
 import { getFoamVsCodeConfig } from './services/config';
 
 /**
@@ -39,15 +38,6 @@ export function getAttachmentsExtensions() {
     .map(ext => '.' + ext.trim());
 }
 
-export function getWikilinkDefinitionSetting():
-  | 'withExtensions'
-  | 'withoutExtensions'
-  | 'off' {
-  return workspace
-    .getConfiguration('foam.edit')
-    .get('linkReferenceDefinitions', 'withoutExtensions');
-}
-
 /** Retrieve the list of file ignoring globs. */
 export function getIgnoredFilesSetting(): GlobPattern[] {
   return [
@@ -55,36 +45,4 @@ export function getIgnoredFilesSetting(): GlobPattern[] {
     ...workspace.getConfiguration().get('foam.files.ignore', []),
     ...Object.keys(workspace.getConfiguration().get('files.exclude', {})),
   ];
-}
-
-/** Retrieves the maximum length for a Graph node title. */
-export function getTitleMaxLength(): number {
-  return workspace.getConfiguration('foam.graph').get('titleMaxLength');
-}
-
-/** Retrieve the graph's style object */
-export function getGraphStyle(): object {
-  return workspace.getConfiguration('foam.graph').get('style');
-}
-
-export function getFoamLoggerLevel(): LogLevel {
-  return workspace.getConfiguration('foam.logging').get('level') ?? 'info';
-}
-
-/** Retrieve the orphans configuration */
-export function getOrphansConfig(): GroupedResourcesConfig {
-  const orphansConfig = workspace.getConfiguration('foam.orphans');
-  const exclude: string[] = orphansConfig.get('exclude');
-  return { exclude };
-}
-
-/** Retrieve the placeholders configuration */
-export function getPlaceholdersConfig(): GroupedResourcesConfig {
-  const placeholderCfg = workspace.getConfiguration('foam.placeholders');
-  const exclude: string[] = placeholderCfg.get('exclude');
-  return { exclude };
-}
-
-export interface GroupedResourcesConfig {
-  exclude: string[];
 }
